+++ src/quasinewton.jl
@@ -13,7 +13,7 @@
 #
 
 # Improvements:
-# - skip updates such that rho <= 0;
+# - skip updates such that rho ≤ 0;
 # - check for sufficient descent condition and implement restarts;
 # - add other convergence criteria;
 # - in case of early stop, revert to best solution so far;
@@ -27,6 +27,7 @@ export
     vmlmb_CUTEst
 
 # Imports from other packages.
+using LinearAlgebra
 using Printf
 using TypeUtils
 using NumOptBase:
@@ -39,13 +40,68 @@ using NumOptBase:
     unblocked_variables!
 
 # Imports from parent module.
-using ..OptimPackNextGen
-using ..OptimPackNextGen.LineSearches
-using ..OptimPackNextGen.VectOps
-using ..OptimPackNextGen: auto_differentiate!, copy_variables, get_tolerances
+using  ..OptimPackNextGen
+using  ..OptimPackNextGen.LineSearches
+using  ..OptimPackNextGen.VectOps
+using  ..OptimPackNextGen: auto_differentiate!, copy_variables, get_tolerances
+import ..OptimPackNextGen: get_reason
 
 # Use the same floating point type for scalars as in OptimPackNextGen.
 import OptimPackNextGen.Float
+
+@enum Status begin
+    LINESEARCH_ERROR     = -3
+    TOO_MANY_EVALUATIONS = -2
+    TOO_MANY_ITERATIONS  = -1
+    SEARCHING            =  0
+    CONVERGENCE_X        =  1
+    CONVERGENCE_FATOL    =  2
+    CONVERGENCE_FRTOL    =  3
+    CONVERGENCE_FMIN     =  4
+    CONVERGENCE_G        =  5
+end
+
+"""
+    VMLMB.Stats{F}(; fx::Real, gnorm::Real, step::Real, seconds::Real,
+                     iter::Integer, eval::Integer, rejects::Integer,
+                     status::Status)
+
+yields an immutable object collecting information returned by the VMLMB method.
+Type parameter `F` is the floating-point type for `fx`, `gnorm`, and `step`.
+All members are mandatory and are specified by keyword.
+
+"""
+struct Stats{F<:AbstractFloat}
+    fx::F            # Objective function value.
+    gnorm::F         # Euclidean norm of the (projected) gradient.
+    step::F          # Line-search step length.
+    seconds::Float64 # Execution time in seconds.
+    iter::Int        # Number of iterations.
+    eval::Int        # Number of objective function evaluations.
+    rejects::Int     # Number of rejected updates.
+    status::Status   # Algorithm status.
+    function Stats{F}(; fx::Real, gnorm::Real, step::Real, seconds::Real,
+                        iter::Integer, eval::Integer, rejects::Integer,
+                        status::Status) where {F}
+        return new{F}(fx, gnorm, step, seconds, iter, eval, rejects, status)
+    end
+end
+
+LinearAlgebra.issuccess(stats::Stats) = issuccess(stats.status)
+LinearAlgebra.issuccess(status::Status) = Integer(status) > 0
+
+get_reason(stats::Stats) = get_reason(stats.status)
+get_reason(status::Status) =
+    status == LINESEARCH_ERROR     ? "Error in line-search" :
+    status == TOO_MANY_EVALUATIONS ? "Too many function evaluations" :
+    status == TOO_MANY_ITERATIONS  ? "Too many iterations" :
+    status == SEARCHING            ? "Work in progress" :
+    status == CONVERGENCE_X        ? "Convergence with `x` test satisfied" :
+    status == CONVERGENCE_FATOL    ? "Convergence with `fatol` test satisfied" :
+    status == CONVERGENCE_FRTOL    ? "Convergence with `frtol` test satisfied" :
+    status == CONVERGENCE_FMIN     ? "Convergence with `f(x) ≤ fmin`" :
+    status == CONVERGENCE_G        ? "Convergence with `g` test satisfied" :
+    "Unknown status"
 
 struct VMLMBConfig{FP<:AbstractFloat}
     method::Symbol # one of `:LBFGS`, `:BLMVM`, or `:VMLMB`
@@ -72,22 +128,36 @@ is_bounded(cfg::VMLMBConfig) = ! is_unconstrained(cfg)
 is_bounded(Ω::BoundedSet) = (NumOptBase.is_bounded_below(Ω.lower) |
                              NumOptBase.is_bounded_above(Ω.upper))
 
+const default_xtol = (0.0, 1.0e-7)
+const default_ftol = (0.0, 1.0e-8)
+const default_gtol = (0.0, 1.0e-6)
+const default_mem  = 5
+
 """
-## VMLMB: limited memory BFGS method with optional bounds
+    vmlmb(fg!, x0; kwds...) -> stats, x
 
-    x = vmlmb(fg!, x0; mem=..., lower=..., upper=..., ftol=..., fmin=...)
+attempts to solve the constrained problem:
 
-computes a local minimizer of a function of several variables by a limited
-memory variable metric method.  The caller provides a function `fg!` to compute
-the value and the gradient of the objective function as follows:
+    min f(x)   subject to   x ∈ Ω   with   Ω = { x ∈ ℝⁿ | ℓ ≤ x ≤ u }
+
+by the VMLMB method, a Variable Metric method with Limited Memory and optional
+Bound constraints. Argument `fg!` implements the objective function `f(x)` and
+its gradient `∇f(x)`. Argument `x0 ∈ ℝⁿ` gives an initial approximation of the
+variables (its contents is left unchanged and it does not need to be feasible).
+The result is a 2-tuple `(stats, x)` with `stats` a structure with information
+about the algorithm computations and `x ∈ Ω` which, provided `issuccess(stats)`
+is true, is an approximate local minimizer of the objective function on the
+feasible set `Ω`.
+
+The caller must provide a function `fg!` to compute the value and the gradient
+of the objective function as follows:
 
     fx = fg!(x, gx)
 
-where `x` are the current variables, `fx` is the value of the function at `x`
-and the contents of `gx` has to be overwritten with the gradient at `x` (when
-`fg!` is called, `gx` is already allocated as `gx = vcreate(x0)`).  Argument
-`x0` gives the initial approximation of the variables (its contents is left
-unchanged).  The best solution found so far is returned in `x`.
+where `x` stores the current variables, `fx` is the value of the function at
+`x` and the contents of `gx` has to be overwritten with the gradient at `x`
+(when `fg!` is called, `gx` is already allocated). The best solution found so
+far is returned in `x`.
 
 Another possibility is to specify keyword `autodiff = true` and rely on
 automatic differentiation to compute the gradient:
@@ -100,18 +170,25 @@ argument and returns the value of the objective function:
     fx = f(x)
 
 The method [`OptimPackNextGen.auto_differentiate!`](@ref) is called to compute
-the gradient of the objective function, say `f`.  This method may be extended
-for the specific type of `f`.  An implementation of `auto_differentiate!` is
+the gradient of the objective function, say `f`. This method may be extended
+for the specific type of `f`. An implementation of `auto_differentiate!` is
 provided by `OptimPackNextGen` if the `Zygote` package is loaded.
 
 The following keywords are available:
 
-* `mem` specifies the amount of storage.
+* `mem` specifies the amount of storage as the number of previous steps
+  memorized to build an L-BFGS model of the inverse Hessian of the objective
+  function. By default `mem = $default_mem`.
+
+* `lower` and `upper` specify the lower and upper bounds for the variables. A
+   bound can be a scalar to indicate that all variables have the same bound
+   value. If the lower (resp. upper) bound is unspecified or set to `±∞`, the
+   variables are assumed to be unbounded below (resp. above). If no bounds are
+   set, VMLMB amounts to an unconstrained limited memory BFGS method (L-BFGS).
 
 * `autodiff` is a boolean specifying whether to rely on automatic
-  differentiation by calling [`OptimPackNextGen.auto_differentiate!](@ref).
-  If not specified, this keyword is assumed to be `false`.
-  You may use:
+  differentiation by calling [`OptimPackNextGen.auto_differentiate!](@ref). If
+  not specified, this keyword is assumed to be `false`. You may use:
 
       autodiff = !applicable(fg!, x0, x0)
 
@@ -120,65 +197,74 @@ The following keywords are available:
 * `xtol` is a tuple of two nonnegative reals specifying respectively the
   absolute and relative tolerances for deciding convergence on the variables.
   Convergence occurs if the Euclidean norm of the the difference between
-  successive iterates is less or equal `max(xtol[1], xtol[2]*vnorm2(x))`.  By
-  default, `xtol = (0.0,1e-7)`.
+  successive iterates is less or equal `max(xtol[1], xtol[2]*vnorm2(x))`. By
+  default, `xtol = $default_xtol`.
 
 * `ftol` is a tuple of two nonnegative reals specifying respectively the
-  absolute and relative errors desired in the function.  Convergence occurs if
+  absolute and relative errors desired in the function. Convergence occurs if
   the absolute error between `f(x)` and `f(xsol)` is less than `ftol[1]` or if
   the estimate of the relative error between `f(x)` and `f(xsol)`, where `xsol`
-  is a local minimizer, is less than `ftol[2]`.  By default, `ftol =
-  (0.0,1e-8)`.
+  is a local minimizer, is less than `ftol[2]`. By default, `ftol =
+  $default_ftol`.
 
 * `gtol` is a tuple of two nonnegative reals specifying the absolute and a
   relative thresholds for the norm of the gradient, convergence is assumed as
   soon as:
 
-      ||g(x)|| <= hypot(gtol[1], gtol[2]*||g(x0)||)
+      ||g(x)|| ≤ hypot(gtol[1], gtol[2]*||g(x0)||)
 
   where `||g(x)||` is the Euclidean norm of the gradient at the current
   solution `x`, `||g(x0)||` is the Euclidean norm of the gradient at the
-  starting point `x0`.  By default, `gtol = (0.0,1e-6)`.
+  starting point `x0`. By default, `gtol = $default_gtol`.
 
-* `fmin` specifies a lower bound for the function.  If provided, `fmin` is used
-  to estimate the steepest desxecnt step length this value.  The algorithm
-  exits with a warning if `f(x) < fmin`.
+* `fmin` specifies a lower bound for the function. If provided, `fmin` is used
+  to estimate the steepest desxecnt step length this value. The algorithm exits
+  with a warning if `f(x) < fmin`.
 
 * `maxiter` specifies the maximum number of iterations.
 
 * `maxeval` specifies the maximum number of calls to `fg!`.
 
 * `verb` specifies the verbosity level. It can be a boolean to specify whether
-  to print information at every iteration or an integer to print information
-  every `verb` iteration(s).  No information is printed if `verb` is less or
+  to call the observer at every iteration or an integer to call the observer
+  every `verb` iteration(s). The observer is never called if `verb` is less or
   equal zero. The default is `verb = false`.
 
-* `printer` can be set with a user defined function to print iteration
-  information, its signature is:
+* `observer` can be set with a callable object to print iteration information
+  or inspect the current iterate, its signature is:
 
-      printer(io::IO, iter::Integer, eval::Integer, rejects::Integer,
-              f::Real, gnorm::Real, stp::Real)
+      observer(output::IO, stats, x)
 
-  where `io` is the output stream, `iter` the iteration number (`iter = 0` for
-  the starting point), `eval` is the number of calls to `fg!`, `rejects` is the
-  number of times the computed direction was rejected, `f` and `gnorm` are the
-  value of the function and norm of the gradient at the current point, `stp` is
-  the length of the step to the current point.
+  where `output` is the output stream, `stats` collects information about the
+  current iteration (see below), and `x` is the current iterate.
 
 * `output` specifies the output stream for printing information (`stdout` is
   used by default).
 
-* `lnsrch` specifies the method to use for line searches (the default
-   line search is `MoreThuenteLineSearch`).
-
-* `lower` and `upper` specify the lower and upper bounds for the variables.
-   The bound can be a scalar to indicate that all variables have the same bound
-   value.  If the lower (resp. upper) bound is unspecified or set to `±∞`, the
-   variables are assumed to be unbounded below (resp. above).  If no bounds are
-   set, VMLMB amounts to an unconstrained limited memory BFGS method (L-BFGS).
+* `lnsrch` specifies the method to use for line searches (the default line
+   search is `MoreThuenteLineSearch`).
 
 * `blmvm` can be set true to emulate the BLMVM algorithm of Benson and Moré.
   This option has no effects for an uncostrained problem.
+
+The `stats` object has the following properties:
+
+* `stats.fx` is the objective function value at `x`.
+
+* `stats.gnorm` is the norm of the gradient of the objective function at `x`.
+
+* `stats.step` is the length of the line-search step to the current point.
+
+* `stats.seconds` is the execution time in seconds.
+
+* `stats.iter` is the number of iterations, `0` for the starting point.
+
+* `stats.eval` is the number of function (and gradient) evaluations.
+
+* `stats.rejects` is the number of times the computed direction was
+  rejected and the L-BFGS recursion restarted.
+
+* `stats.status` indicates the status of the algorithm.
 
 
 ### History
@@ -189,14 +275,14 @@ Julia implementation of the original method (Thiébaut, 2002) with some
 improvements and the capability to emulate L-BFGS and BLMVM methods.
 
 The limited memory BFGS method (L-BFGS) was first described by Nocedal (1980)
-who dubbed it SQN.  The method is implemented in MINPACK-2 (1995) by the
-FORTRAN routine VMLM.  The numerical performances of L-BFGS have been studied
-by Liu and Nocedal (1989) who proved that it is globally convergent for
-unfiformly convex problems with a R-linear rate of convergence.  They provided
-the FORTRAN code LBFGS.  The BLMVM and VMLMB algorithms were proposed by Benson
-and Moré (2001) and Thiébaut (2002) to account for separable bound constraints
-on the variables.  These two latter methods are rather different than L-BFGS-B
-by Byrd at al. (1995) which has more overheads and is slower.
+who dubbed it SQN. The method is implemented in MINPACK-2 (1995) by the FORTRAN
+routine VMLM. The numerical performances of L-BFGS have been studied by Liu and
+Nocedal (1989) who proved that it is globally convergent for unfiformly convex
+problems with a R-linear rate of convergence. They provided the FORTRAN code
+LBFGS. The BLMVM and VMLMB algorithms were proposed by Benson and Moré (2001)
+and Thiébaut (2002) to account for separable bound constraints on the
+variables. These two latter methods are rather different than L-BFGS-B by Byrd
+at al. (1995) which has more overheads and is slower.
 
 * J. Nocedal, "*Updating Quasi-Newton Matrices with Limited Storage*" in
   Mathematics of Computation, vol. 35, pp. 773-782 (1980).
@@ -216,10 +302,13 @@ by Byrd at al. (1995) which has more overheads and is slower.
   Astronomical Data Analysis II, Proc. SPIE 4847, pp. 174-183 (2002).
 
 """
-vmlmb(fg!, x0; kwds...) = vmlmb!(fg!, copy_variables(x0); kwds...)
+function vmlmb(fg!, x0; kwds...)
+    x = copy_variables(x0)
+    return vmlmb!(fg!, x; kwds...), x
+end
 
 """
-    vmlmb_CUTEst(name; kwds...) -> x
+    vmlmb_CUTEst(name; kwds...) -> x, stats
 
 yields the solution to the `CUTEst` problem `name` by the VMLMB method. This
 require to have loaded the `CUTest` package.
@@ -229,9 +318,9 @@ vmlmb_CUTEst(arg...; kwds...) =
     error("invalid arguments or `CUTEst` package not yet loaded")
 
 """
-     vmlmb!(fg!, x; mem=..., lower=..., upper=..., ftol=..., fmin=...) -> x
+     vmlmb!(fg!, x; mem=..., lower=..., upper=..., ftol=..., fmin=...) -> stats
 
-finds a local minimizer of `f(x)` starting at `x` and stores the best solution
+finds a local minimizer of `f(x)` starting at `x` and storing the best solution
 in `x`. Method `vmlmb!` is the in-place version of `vmlmb` (which to see).
 
 """
@@ -244,12 +333,12 @@ function vmlmb!(fg!, x::AbstractArray{T,N};
                 fmin::Real = -Inf,
                 maxiter::Integer = typemax(Int),
                 maxeval::Integer = typemax(Int),
-                xtol::NTuple{2,Real} = (0.0, 1e-7),
-                ftol::NTuple{2,Real} = (0.0, 1e-8),
-                gtol::NTuple{2,Real} = (0.0, 1e-6),
+                xtol::NTuple{2,Real} = default_xtol,
+                ftol::NTuple{2,Real} = default_ftol,
+                gtol::NTuple{2,Real} = default_gtol,
                 epsilon::Real = 0.0,
                 verb::Integer = false,
-                printer::Function = print_iteration,
+                observer = default_observer,
                 output::IO = stdout,
                 lnsrch::Union{LineSearch{Float},Nothing} = nothing) where {T<:AbstractFloat,N}
     # Check settings.
@@ -278,33 +367,32 @@ function vmlmb!(fg!, x::AbstractArray{T,N};
         autodiff)
 
     # Call the real method.
-    _vmlmb!(fg!, x, Ω, cfg, lnsrch, printer, output)
+    return _vmlmb!(fg!, x, Ω, cfg, lnsrch, observer, output)
 end
 
 # Provide a default line search method if needed.
 function _vmlmb!(fg!, x::AbstractArray, Ω::BoundedSet, cfg::VMLMBConfig,
-                 lnsrch::Nothing, printer::Function, output::IO)
+                 lnsrch::Nothing, observer, output::IO)
     if is_unconstrained(cfg)
-        ls = MoreThuenteLineSearch(Float; ftol=1e-3, gtol=0.9, xtol=0.1)
-        return _vmlmb!(fg!, x, Ω, cfg, ls, printer, output)
+        ls = MoreThuenteLineSearch{Float}(ftol=1e-3, gtol=0.9, xtol=0.1)
+        return _vmlmb!(fg!, x, Ω, cfg, ls, observer, output)
     else
-        ls = MoreToraldoLineSearch(Float; ftol=1e-3, gamma=(0.1,0.5))
-        return _vmlmb!(fg!, x, Ω, cfg, ls, printer, output)
+        ls = MoreToraldoLineSearch{Float}(ftol=1e-3, gamma=(0.1,0.5))
+        return _vmlmb!(fg!, x, Ω, cfg, ls, observer, output)
     end
 end
 
 # The real worker.
 function _vmlmb!(fg!, x::T, Ω::BoundedSet,
                  cfg::VMLMBConfig, lnsrch::LineSearch{FP},
-                 printer::Function, output::IO) where {T<:AbstractArray,
-                                                       FP<:AbstractFloat}
-
+                 observer, output::IO) where {T<:AbstractArray,
+                                              FP<:AbstractFloat}
+    t0 = time()
     STPMIN = FP(1e-20) # FIXME: should be in config
     STPMAX = FP(1e+20) # FIXME: should be in config
 
     act = similar(x) # FIXME: not used if not bounded
     mem = min(cfg.mem, length(x))
-    reason = "" # FIXME: use enumeration
     fminset::Bool = (! isnan(cfg.fmin) && cfg.fmin > -Inf) # FIXME: simplify
 
     mark = 1    # index of most recent step and gradient difference
@@ -357,6 +445,7 @@ function _vmlmb!(fg!, x::T, Ω::BoundedSet,
     stage = 0
 
     bounded = is_bounded(cfg)
+    status = SEARCHING
     while true
 
         # Compute value of function and gradient, register best solution so far
@@ -383,21 +472,23 @@ function _vmlmb!(fg!, x::T, Ω::BoundedSet,
             end
             if gnorm ≤ gtest
                 stage = 3
-                if gnorm == 0
-                    reason = "a stationary point has been found"
-                elseif is_bounded(cfg)
-                    reason = "projected gradient norm sufficiently small"
-                else
-                    reason = "gradient norm sufficiently small"
-                end
+                status = CONVERGENCE_G
+                # if gnorm == 0
+                #     status = CONVERGENCE_KKT
+                #     reason = "a stationary point has been found"
+                # elseif is_bounded(cfg)
+                #     reason = "projected gradient norm sufficiently small"
+                # else
+                #     reason = "gradient norm sufficiently small"
+                # end
                 if eval > 1
                     iter += 1 # FIXME: do this here???
                 end
             end
         end
-        if fminset && f < cfg.fmin # FIXME: do not do that and use ≤
+        if f ≤ cfg.fmin
             stage = 4
-            reason = "f < fmin"
+            status = CONVERGENCE_FMIN
         end
 
         if stage == 1
@@ -405,19 +496,19 @@ function _vmlmb!(fg!, x::T, Ω::BoundedSet,
             vcombine!(s, 1, x, -1, S[mark])
 
             # Line search is in progress.
-            if usederivatives(lnsrch)
+            if use_derivatives(lnsrch)
                 if is_bounded(cfg)
                     gd = vdot(FP, g, s)/stp
                 else
                     gd = -vdot(FP, g, d)
                 end
             end
-            task = iterate!(lnsrch, stp, f, gd)
-            if task == :SEARCH
-                stp = getstep(lnsrch)
-            elseif task == :CONVERGENCE
-                # Line search has converged.  Increment iteration counter
-                # and check for stopping conditions.
+            state = iterate!(lnsrch, f, gd)
+            if state == :SEARCHING
+                stp = get_step(lnsrch)
+            elseif state == :CONVERGENCE || state == :WARNING
+                # Line search has converged. Increment iteration counter and
+                # check for stopping conditions.
                 iter += 1
                 xtest = max(cfg.xatol, zero(FP))
                 if cfg.xrtol > 0
@@ -425,34 +516,34 @@ function _vmlmb!(fg!, x::T, Ω::BoundedSet,
                 end
                 if vnorm2(FP, s) ≤ xtest
                     stage = 3
-                    reason = "X test satisfied"
+                    status = CONVERGENCE_X
                 else
                     delta = max(abs(f - f0), stp*abs(gd0))
                     if delta ≤ cfg.fatol
                         stage = 3
-                        reason = "fatol test satisfied"
+                        status = CONVERGENCE_FATOL
                     elseif delta ≤ cfg.frtol*abs(f0)
                         stage = 3
-                        reason = "frtol test satisfied"
+                        status = CONVERGENCE_FRTOL
                     elseif iter ≥ cfg.maxiter
                         stage = 4
-                        reason = "too many iterations"
+                        status = TOO_MANY_ITERATIONS
                     elseif eval ≥ cfg.maxeval
                         stage = 4
-                        reason = "too many evaluations"
+                        status = TOO_MANY_EVALUATIONS
                     else
                         stage = 2
                     end
                 end
             else
-                # Line seach terminated with a warningor an error.
+                # Line seach terminated on error.
                 stage = 4
-                reason = getreason(lnsrch)
+                status = LINESEARCH_ERROR # FIXME throw an error
             end
         end
         if stage < 2 && eval ≥ cfg.maxeval
             stage = 4
-            reason = "too many evaluations"
+            status = TOO_MANY_ITERATIONS
         end
 
         if stage != 1
@@ -477,7 +568,9 @@ function _vmlmb!(fg!, x::T, Ω::BoundedSet,
             # Print some information if requested and terminate algorithm if
             # stage ≥ 3.
             if verbose(cfg.verb, iter)
-                printer(output, iter, eval, rejects, f, gnorm, stp)
+                stats = Stats{FP}(; fx = f, gnorm, step = stp, seconds = time() - t0,
+                                  eval, iter, rejects, status)
+                observer(output, stats, x)
             end
             if stage ≥ 3
                 break
@@ -563,11 +656,11 @@ function _vmlmb!(fg!, x::T, Ω::BoundedSet,
             end
 
             # Initialize the line search.
-            task = start!(lnsrch, f0, gd0, stp; stpmin=stpmin, stpmax=stpmax)
-            if task != :SEARCH
+            state = start!(lnsrch, f0, gd0, stp; stpmin=stpmin, stpmax=stpmax)
+            if state != :SEARCHING
                 # Something wrong happens.
                 stage = 4
-                reason = getreason(lnsrch)
+                status = lnsrch.status # FIXME: use a better scale
                 break
             end
             stage = 1 # line search is in progress
@@ -583,11 +676,11 @@ function _vmlmb!(fg!, x::T, Ω::BoundedSet,
     if verbose(cfg.verb, 0) #always print last line if verb>0
         color = (stage > 3 ? :red : :green)
         prefix = (stage > 3 ? "WARNING: " : "CONVERGENCE: ")
+        reason = get_reason(status)
         printstyled(output, "# ", prefix, reason, "\n"; color=color)
-        #elseif stage > 3
-        #@warn(reason)
     end
-    return x
+    return Stats{FP}(; fx = bestf, gnorm = bestgnorm, step = beststp,
+                     seconds = time() - t0, eval, iter, rejects, status)
 end
 
 """
@@ -617,25 +710,19 @@ yields whether to print information at iteration `iter` with verbose level
 `verb`.
 
 """
-verbose(verb::Integer,iter::Integer) =
-    (verb > 0 && iter%verb == 0)
+verbose(verb::Integer, iter::Integer) = (verb > 0 && iter%verb == 0)
 
-function print_iteration(iter::Integer, eval::Integer, rejects::Integer,
-                         f::Real, gnorm::Real, step::Real)
-    print_iteration(stdout, iter, eval, rejects, f, gnorm, step)
-end
-
-function print_iteration(io::IO, iter::Integer, eval::Integer, rejects::Integer,
-                         f::Real, gnorm::Real, step::Real)
-    if iter == 0
-        @printf(io, "#%s%s\n#%s%s\n",
+function default_observer(output::IO, stats::Stats, x::AbstractArray)
+    if stats.iter == 0
+        @printf(output, "#%s%s\n#%s%s\n",
                 " ITER   EVAL   REJECTS",
                 "          F(X)           ||G(X)||    STEP",
                 "----------------------",
                 "-------------------------------------------")
     end
-    @printf(io, " %5d  %5d  %5d  %24.16E %9.2E %9.2E\n",
-            iter, eval, rejects, f, gnorm, step)
+    @printf(output, " %5d  %5d  %5d  %24.16E %9.2E %9.2E\n",
+            stats.iter, stats.eval, stats.rejects, stats.fx, stats.gnorm,
+            stats.step)
 end
 
 #------------------------------------------------------------------------------
@@ -648,9 +735,9 @@ One of the calls:
     modif = apply_lbfgs!(S, Y, rho, H0!,   m, mark, v, alpha)
 
 computes the matrix-vector product `H*v` where `H` is the limited memory
-inverse BFGS approximation.  Operation is done *in-place*: on entry, argument
+inverse BFGS approximation. Operation is done *in-place*: on entry, argument
 `v` contains the input vector `v`; on exit, argument `v` contains the
-matrix-vector product `H*v`.  The returned value, `modif`, is a boolean
+matrix-vector product `H*v`. The returned value, `modif`, is a boolean
 indicating whether the initial vector was modified (BFGS updates are skipped if
 `rho[k] ≤ 0`).
 
@@ -660,7 +747,7 @@ The most recent step and gradient difference are stored in `S[mark]` and
 `Y[mark]`, respectively.
 
 Argument `rho` contains the inner products of the steps and the gradient
-differences: `rho[k] = vdot(S[k],Y[k])`.  On exit rho is unchanged.
+differences: `rho[k] = vdot(S[k],Y[k])`. On exit rho is unchanged.
 
 Argument `alpha` is a work vector of length at least `m`.